+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Demos/LiveFinance.cs
@@ -69,7 +69,7 @@ public partial class LiveFinance : Form, IDemoWindow
             OHLCs[lastOhlcIndex] = updatedOhlc;
         }
 
-        formsPlot1.Plot.Axes.AutoScale();
+        formsPlot1.Plot.Axes.AutoScaleExpand();
         formsPlot1.Refresh();
     }
 }