+++ src/ScottPlot5/ScottPlot5/Primitives/AxisLimits.cs
@@ -65,6 +65,13 @@ public readonly struct AxisLimits : IEquatable<AxisLimits>
     public AxisLimits InvertedVertically() => new(Left, Right, Top, Bottom);
     public AxisLimits InvertedHorizontally() => new(Right, Left, Bottom, Top);
 
+    public AxisLimits ExpandedToInclude(AxisLimits otherLimits)
+    {
+        ExpandingAxisLimits expandingLimits = new(this);
+        expandingLimits.Expand(otherLimits);
+        return expandingLimits.AxisLimits;
+    }
+
     public static AxisLimits FromPoint(double x, double y)
     {
         return new AxisLimits(x, x, y, y);