+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * SignalXY: Support vertical orientation (#3253) _Thanks @manaruto_
 * Data logger and streamer: The property `Data` has been renamed to `DataSource` (#3260)
 * SignalConst: The property `Data` has been renamed to `DataSource` (#3260)
+* Axes: Added `AutoScaleExpand()` to zoom out to fit data only if necessary (#3259)
 
 ## ScottPlot 4.1.71 (not yet on NuGet)
 * Assembly: All ScottPlot packages use the same strong name signing key (#3235, #3241) _Thanks @mlessmann_